typescript refactoring
